Share exponential backoff code and fix logic for delete task failure
